fix: change content in page and layout files
